requirements.txt
Python 3.8+
Librerias:
    pycryptodome